moved some stuff to be correctly listed/useable as transformers
