+++ controllers/index.js
@@ -1,0 +1,7 @@
+/**
+ * Controllers
+ */
+module.exports.lesson = require('./lessonController');
+module.exports.course = require('./courseController');
+module.exports.path = require('./pathController');
+module.exports.user = require('./user');
