+++ db/seed.js
@@ -1,47 +1,7 @@
 /* eslint-disable no-console */
-require('dotenv/config');
-const axios = require('axios');
 const { sequelize } = require('../config/config');
+const { transformLesson } = require('../utils/transformers');
 const { Path, Course, Lesson } = require('../models');
-
-const YOUTUBE_API_URL = 'https://www.googleapis.com/youtube/v3/';
-
-/**
- * Takes in a string in the YouTube ISO-8601 duration format and parses it into a single
- * integer representing the total number of seconds for the duration
- * @param {string} duration the YouTube string format for duration to be parsed
- * @returns {number} the total number of seconds
- */
-function parseDuration(duration) {
-  // YouTube uses a slightly modified ISO-8601 duration format
-  // (PnYnMnWnDTnHnMnS) which includes weeks along with the normal other units
-  const numbers = '\\d+(?:[\\.,]\\d+)?';
-  const period_pattern = `P(?<years>${numbers}Y)?(?<months>${numbers}M)?(?<weeks>${numbers}W)?(?<days>${numbers}D)?`;
-  const time_pattern = `T(?<hours>${numbers}H)?(?<minutes>${numbers}M)?(?<seconds>${numbers}S)?`;
-  const iso8601 = new RegExp(`${period_pattern}${time_pattern}`);
-  const { groups } = duration.match(iso8601);
-  // coefficients to convert a time unit to the next smaller time unit
-  const reduce_unit_coeffs = {
-    years: 12,
-    months: 4,
-    weeks: 7,
-    days: 24,
-    hours: 60,
-    minutes: 60,
-    seconds: 1,
-  };
-  // Each step of this reduce takes the total number of the current time units
-  // and converts it to the next smaller time unit. Since the smallest unit is seconds,
-  // the end result will be the total number of seconds all together.
-  return Object.keys(groups).reduce((curr_unit_total, time_unit) => {
-    let num = 0;
-    if (groups[time_unit] !== undefined) {
-      // removes the unit character from the end and parses
-      num = parseInt(groups[time_unit].slice(0, -1), 10);
-    }
-    return (curr_unit_total + num) * reduce_unit_coeffs[time_unit];
-  }, 0);
-}
 
 /**
  * An object containing the manually created parts of lesson data
@@ -63,33 +23,11 @@ function parseDuration(duration) {
  * @param {string} course_id the uuid of the course this lesson belongs to
  * @returns an object with the data of the newly created/seeded lesson
  */
-async function seedLesson(
-  { slug, title, description, order, start, end, video_id },
-  course_id,
-) {
+async function seedLesson(lesson_data, course_id) {
   // TODO: error handling/graceful failure
-  const { data: youtube_data } = await axios.get(
-    `${YOUTUBE_API_URL}videos?part=snippet,contentDetails&id=${video_id}&key=${process.env.GOOGLE_API_KEY}`,
+  const { dataValues: new_lesson } = await Lesson.create(
+    await transformLesson(lesson_data, course_id),
   );
-  const { snippet, contentDetails } = youtube_data.items[0];
-  const { dataValues: new_lesson } = await Lesson.create({
-    slug,
-    title,
-    description,
-    order,
-    start,
-    end,
-    video_id,
-    length:
-      start !== undefined && end !== undefined
-        ? end - start
-        : parseDuration(contentDetails.duration),
-    video_title: snippet.title,
-    video_description: snippet.description,
-    channel_id: snippet.channelId,
-    channel_name: snippet.channelTitle,
-    courseUuid: course_id,
-  });
   console.log(`Seeded lesson '${new_lesson.title}'`);
   return new_lesson;
 }
@@ -111,7 +49,7 @@ async function seedLesson(
  * data model to create a new table entry for this object
  * @param {CourseData} course_data the data for the course to be seeded
  * @param {string} path_id the uuid of the path this course belongs to
- * @returns an object with the data of the newly created/seeded course
+ * @returns {CourseData} an object with the data of the newly created/seeded course
  */
 async function seedCourse({ slug, title, description, image_name, order }, path_id) {
   // TODO: error handling/graceful failure
@@ -140,7 +78,7 @@ async function seedCourse({ slug, title, description, image_name, order }, path_
  * Takes in some manually created path data, constructs a new path object using this
  * data, and uses the Path data model to create a new table entry for this object
  * @param {PathData} path_data the data for the path to be seeded
- * @returns an object with the data of the newly created/seeded path
+ * @returns {PathData} an object with the data of the newly created/seeded path
  */
 async function seedPath({ title, subtitle, image_name }) {
   // TODO: error handling/graceful failure
